+++ datumaro/datumaro/plugins/coco_format/converter.py
@@ -12,7 +12,7 @@ from itertools import groupby
 
 import pycocotools.mask as mask_utils
 
-import datumaro.util.annotation_tools as anno_tools
+import datumaro.util.annotation_util as anno_tools
 import datumaro.util.mask_tools as mask_tools
 from datumaro.components.converter import Converter
 from datumaro.components.extractor import (_COORDINATE_ROUNDING_DIGITS,
@@ -202,7 +202,7 @@ class _InstancesConverter(_TaskConverter):
 
         anns = boxes + polygons + masks
         leader = anno_tools.find_group_leader(anns)
-        bbox = anno_tools.compute_bbox(anns)
+        bbox = anno_tools.max_bbox(anns)
         mask = None
         polygons = [p.points for p in polygons]